[Dashboard] Adjusted bot endpoints according to the requirements of the bot

+++ dashboard/server/api/[guildId]/lobbies/[lobbyId]/index.delete.ts
@@ -4,6 +4,20 @@ export default defineEventHandler(async event => {
   const guildId = getRouterParam(event, 'guildId');
   const lobbyId = getRouterParam(event, 'lobbyId');
 
+  const lobby = await prisma.lobby.findFirst({
+    where: {
+      guildId: guildId,
+      entryPointId: lobbyId
+    }
+  });
+
+  if (!lobby) {
+    throw createError({
+      statusCode: 404,
+      statusMessage: 'Lobby not found'
+    });
+  }
+
   await prisma.lobby.delete({
     where: {
       guildId: guildId,
